Fixed menu issue and made menu reversing possible
Previously had an issue where the tabselectionpanel would dissapear when first trying to open any of the next tabs, and then not being able to reopen any menu's for a few seconds. It's been resolved through lots of changes but one of the big bad guys was the key being TAB, which somehow does some UI traversing bs so I changed it to B amongst other changes.